Bug Fixes and Local NinjaOne Credential support

+++ Modules/CippExtensions/NinjaOne/Invoke-NinjaOneTenantSync.ps1
@@ -2253,9 +2253,14 @@ function Invoke-NinjaOneTenantSync {
         $Result = Invoke-WebRequest -uri "https://$($Configuration.Instance)/api/v2/organization/$($MappedTenant.NinjaOne)/custom-fields" -Method PATCH -Headers @{Authorization = "Bearer $($token.access_token)" } -ContentType 'application/json' -Body ($NinjaOrgUpdate | ConvertTo-Json -Depth 100)
 
         Write-Host "Cleaning Users Cache"
-        Remove-AzDataTableEntity @UsersTable -Entity ($ParsedUsers | Select-Object PartitionKey, RowKey)
+        if (($ParsedUsers | Measure-Object).count -gt 0) {
+            Remove-AzDataTableEntity @UsersTable -Entity ($ParsedUsers | Select-Object PartitionKey, RowKey)
+        }
+        
         Write-Host "Cleaning Device Cache"
-        Remove-AzDataTableEntity @DeviceTable -Entity ($ParsedDevices | Select-Object PartitionKey, RowKey)
+        if (($ParsedDevices | Measure-Object).count -gt 0) {
+            Remove-AzDataTableEntity @DeviceTable -Entity ($ParsedDevices | Select-Object PartitionKey, RowKey)
+        }
         
         Write-Host "Total Fetch Time: $((New-TimeSpan -Start $StartTime -End $FetchEnd).TotalSeconds)"
         Write-Host "Completed Total Time: $((New-TimeSpan -Start $StartTime -End (Get-Date)).TotalSeconds)"
